Defer packaged pytest plugin to follow-up
