+++ ml_peg/calcs/conftest.py
@@ -1,0 +1,43 @@
+"""Configure pytest for calculations."""
+
+from __future__ import annotations
+
+from pytest import Config, Parser
+
+from ml_peg import models
+
+
+def pytest_addoption(parser: Parser) -> None:
+    """
+    Add custom CLI inputs to pytest.
+
+    Parameters
+    ----------
+    parser
+        Pytest parser object.
+    """
+    parser.addoption(
+        "--run-mock",
+        action="store_true",
+        default=False,
+        help="Include mock model in tests",
+    )
+    parser.addoption(
+        "--mock-only",
+        action="store_true",
+        default=False,
+        help="Only run mock model, ignoring other models",
+    )
+
+
+def pytest_configure(config: Config) -> None:
+    """
+    Configure pytest to custom CLI inputs.
+
+    Parameters
+    ----------
+    config
+        Pytest configuration object.
+    """
+    models.run_mock = config.getoption("--run-mock")
+    models.mock_only = config.getoption("--mock-only")
