+++ pyproject.toml
@@ -35,7 +35,6 @@ dependencies = [
     "mlipx<0.2,>=0.1.5",
     "molify<0.3,>=0.2.2",
     "openpyxl<4,>=3.1.5",
-    "pytest>=8.0,<10.0",
     "scipy<1.17,>=1",
     "scikit-learn<2,>=1.7.2",
     "tqdm<5,>=4.67.3",
@@ -95,9 +94,6 @@ vivace = [
 [project.scripts]
 ml_peg = "ml_peg.cli.cli:app"
 
-[project.entry-points.pytest11]
-"ml_peg.pytest_plugin" = "ml_peg.pytest_plugin"
-
 [project.urls]
 Repository = "https://github.com/ddmms/ml-peg/"
 Documentation = "https://ddmms.github.io/ml-peg/"
@@ -107,6 +103,7 @@ dev = [
     "coverage[toml]<8.0.0,>=7.4.1",
     "ipykernel>=6.29.5",
     "pgtest<2.0.0,>=1.3.2",
+    "pytest>=8.0,<10.0",
     "pytest-cov<5.0.0,>=4.1.0",
     "tox-uv<2.0,>=1.16.1",
     "wheel<1.0,>=0.42",
